RanEffdata {mixtools}	R Documentation

Simulated Data from 2-Component Mixture of Regressions with Random Effects

Description

This data set was generated from a 2-component mixture of regressions with random effects.

Usage

data(RanEffdata)

Format

This data set consists of a list with 100 25x3 matrices. The first column is the response variable, the second column is a column of 1's and the last column is the predictor variable.

See Also

regmixEM.mixed

[Package mixtools version 2.0.0.1 Index]