depth {mixtools}	R Documentation

Elliptical and Spherical Depth

Description

Computation of spherical or elliptical depth.

Usage

 depth(pts, x, Cx = var(x))

Arguments

pts
A kxd matrix containing the k points that one wants to compute the depth. Each row is a point.

x
A nxd matrix containing the reference data. Each row is an observation.

Cx
A dxd scatter matrix for the data x where the default is var(x). When Cx = I(d), it returns the sphercial depth.

Value

depth returns a k-vector where each entry is the elliptical depth of a point in pts.

Note

depth is used in regcr.

References

Elmore, R. T., Hettmansperger, T. P. and Xuan, F. (2000) Spherical Data Depth and a Multivariate Median, Proceedings of Data Depth: Robust Multivariate Statistical Analysis, Computational Geometry and Applications.

See Also

regcr

Examples

  set.seed(100)
  x <- matrix(rnorm(200),nc = 2)
  depth(x[1:3, ], x)

[Package mixtools version 2.0.0.1 Index]