ellipse {mixtools}	R Documentation

Draw Two-Dimensional Ellipse Based on Mean and Covariance

Description

Draw a two-dimensional ellipse that traces a bivariate normal density contour for a given mean vector, covariance matrix, and probability content.

Usage

ellipse(mu, sigma, alpha = .05, npoints = 250, newplot = FALSE,
        draw = TRUE, ...)

Arguments

mu
A 2-vector giving the mean.

sigma
A 2x2 matrix giving the covariance matrix.

alpha
Probability to be excluded from the ellipse. The default value is alpha = .05, which results in a 95% ellipse.

npoints
Number of points comprising the border of the ellipse.

newplot
If newplot = TRUE and draw = TRUE, plot the ellipse on a new plot. If newplot = FALSE and draw = TRUE, add the ellipse to an existing plot.

draw
If TRUE, draw the ellipse.

...
Graphical parameters passed to lines or plot command.

Value

ellipse returns an npointsx2 matrix of the points forming the border of the ellipse.

References

Johnson, R. A. and Wichern, D. W. (2002) Applied Multivariate Statistical Analysis, Fifth Edition, Prentice Hall.

See Also

regcr

Examples

## Produce a 95% ellipse with the specified mean and covariance structure. 

mu <- c(1, 3)
sigma <- matrix(c(1, .3, .3, 1.5), 2, 2)

ellipse(mu, sigma, npoints = 200, newplot = TRUE)
 

[Package mixtools version 2.0.0.1 Index]