lambda {mixtools}	R Documentation

Local Estimation for Lambda in Mixtures of Regressions

Description

Return local estimates of the mixing proportions from each component of a mixture of regressions model using output from an EM algorithm.

Usage

lambda(z, x, xi, h = NULL, kernel = c("Gaussian", "Beta", 
       "Triangle", "Cosinus", "Optcosinus"), g = 0)

Arguments

z
An nxk matrix of posterior probabilities obtained from the EM algorithm.

x
A vector of values for which the local estimation is calculated.

xi
An nx(p-1) matrix of the predictor values.

h
The bandwidth controlling the size of the window used for the local estimation.

kernel
The type of kernel to be used for the local estimation.

g
A shape parameter required for the symmetric beta kernel. The default is g = 0 which yields the uniform kernel. Some common values are g = 1 for the Epanechnikov kernel, g = 2 for the biweight kernel, and g = 3 for the triweight kernel.

Value

lambda returns local estimates of the mixing proportions for the inputted x vector.

Note

lambda is for use within regmixEM.loc.

See Also

regmixEM.loc

[Package mixtools version 2.0.0.1 Index]