ldmult {mixtools}	R Documentation

Log-Density for Multinomial Distribution

Description

Return the logarithm of the multinomial density function.

Usage

ldmult(y, theta)

Arguments

y
A vector of multinomial counts.

theta
A vector of multinomial probabilities. May have same number of components as or one fewer component than y. In the latter case, an extra component is appended so that theta sums to one.

Details

This function is called by multmixEM.

Value

ldmult returns the logarithm of the multinomial density with parameter theta, evaluated at y.

See Also

multmixEM

Examples

y <- c(2, 2, 10)
theta <- c(0.2, 0.3, 0.5)
ldmult(y, theta)

[Package mixtools version 2.0.0.1 Index]