mixtools-internal {mixtools}	R Documentation

Internal 'mixtools' Functions

Description

Internal kernel, semiparametric-related, and miscellaneous functions for the package mixtools.

Usage

dexpmixt(t, lam, rate)
HRkde(cpd, u = cpd[,1], kernelft = triang_wkde, 
      bw = rep(bw.nrd0(as.vector(cpd[,1])), length(cpd[,1])))
inv.logit(eta)
kern.B(x, xi, h, g = 0)
kern.C(x, xi, h)
kern.G(x, xi, h)
kern.O(x, xi, h)
kern.T(x, xi, h)
kfoldCV(h, x, nbsets = 2, w = rep(1, length(x)), 
        lower = mean(x) - 5*sd(x), upper = mean(x) + 5*sd(x))
KMintegrate(s) 
KMod(cpd, already.ordered = TRUE)
ldc(data, class, score)
logit(mu)
npMSL_old(x, mu0, blockid = 1:ncol(x),
          bw=bw.nrd0(as.vector(as.matrix(x))), samebw = TRUE,
          h=bw, eps=1e-8, maxiter=500, bwiter = maxiter,
          ngrid = 200, post = NULL, verb = TRUE)
plotseq(x, ...)
rlnormscalemix(n, lambda=1, meanlog=1, sdlog=1, scale=0.1)
splitsample(n, nbsets = 2)
triang_wkde(t, u=t, w=rep(1/length(t),length(t)), bw=rep(bw.nrd0(t), length(t)))
wbw.kCV(x, nbfold = 5, w = rep(1, length(x)), 
        hmin = 0.1*hmax, hmax = NULL)

Arguments

x
A vector of values to which local modeling techniques are applied.

xi
An n-vector of data values.

h
The bandwidth controlling the size of the window used for the local estimation around x. This pertains to its usage in the kernel functionns kern.B, kern.C, kern.G, kern.O, and kern.T. For its usage in the kfoldCV function, see updated arguments in the npMSL function.

g
A shape parameter required for the symmetric beta kernel. The default is g = 0 which yields the uniform kernel. Some common values are g = 1 for the Epanechnikov kernel, g = 2 for the biweight kernel, and g = 3 for the triweight kernel.

mu0
See updated arguments in the npMSL function.

blockid
See updated arguments in the npMSL function.

bw
See updated arguments in the npMSL function.

samebw
See updated arguments in the npMSL function.

eps
See updated arguments in the npMSL function.

maxiter
See updated arguments in the npMSL function.

bwiter
See updated arguments in the npMSL function.

ngrid
See updated arguments in the npMSL function.

post
See updated arguments in the npMSL function.

verb
See updated arguments in the npMSL function.

n
See updated arguments in the npMSL function.

nbsets
See updated arguments in the npMSL function.

w
See updated arguments in the npMSL function.

lower
See updated arguments in the npMSL function.

upper
See updated arguments in the npMSL function.

nbfold
See updated arguments in the npMSL function.

hmin
See updated arguments in the npMSL function.

hmax
See updated arguments in the npMSL function.

data
Data, possibly multivariate, fed to the mixturegram function.

class
The number of classes, inputted based on number of components in the mixturegram function.

score
The score vector from LDA used in constructing a mixturegram.

lam
A vector of mixture proportions, should sum to one.

rate
A vector of mixture component rates.

t
Argument for dexpmixt.

mu
A proportion for which to calculate the logit function; i.e., log(mu / (1 - mu)).

eta
Any real value for which to calculate the inverse logit function; i.e., 1 / (1 + exp(eta)).

cpd
Argument for HRkde.

kernelft
Argument for HRkde.

s
Argument for KMintegrate.

meanlog
Argument for rlnormscalemix.

sdlog
Argument for rlnormscalemix.

Details

These are usually not to be called by the user.

See Also

npMSL

[Package mixtools version 2.0.0.1 Index]