perm {mixtools}	R Documentation

Permutation Function

Description

Enumerates the possible permutations of a specified size from the elements of a vector having the same size.

Usage

perm(n, r, v = 1:n)

Arguments

n
Size of the source vector.

r
Size of the target vector.

v
Source vector. Must be a vector of length n.

Details

This function is called by segregmixEM and the associated internal functions. This is also a simplified version of the function permutations found in the package gtools.

Value

perm returns a matrix where each row contains one of the permutations of length r.

See Also

segregmixEM

Examples

perm(3, 3, 2:4)

[Package mixtools version 2.0.0.1 Index]