plot.mvnpEM {mixtools}	R Documentation

Plots of Marginal Density Estimates from the mvnpEM Algorithm Output

Description

Takes an object of class mvnpEM, as the one returned by the mvnpEM algorithm, and returns a set of plots of the density estimates for each coordinate within each multivariate block. All the components are displayed on each plot so it is possible to see the mixture structure for each coordinate and block. The final bandwidth values are also displayed, in a format depending on the bandwidth strategy .

Usage

## S3 method for class 'mvnpEM'
plot(x, truenorm = FALSE, lambda = NULL, mu = NULL, v = NULL, 
            lgdcex = 1, ...)

Arguments

x
An object of class mvnpEM such as the output of the mvnpEM function

truenorm
Mostly for checking purpose, if the nonparametric model is to be compared with a multivariate Gaussian mixture as the true model.

lambda
true weight parameters, for Gaussian models only (see above)

mu
true mean parameters, for Gaussian models only (see above)

v
true covariance matrices, for Gaussian models only (see above)

lgdcex
Character expansion factor for legend.

...
Any remaining arguments are passed to hist.

Value

plot.mvnpEM currently just plots the figure.

See Also

mvnpEM, npEM, density.npEM

Examples

 
# example as in Chauveau and Hoang (2015) with 6 coordinates
## Not run: 
m=2; r=6; blockid <-c(1,1,2,2,3,3) # 3 bivariate blocks 
# generate some data x ...
a <- mvnpEM(x, mu0=2, blockid, samebw=F) # adaptive bandwidth
plot(a) # this S3 method produces 6 plots of univariate marginals
summary(a)
## End(Not run)

[Package mixtools version 2.0.0.1 Index]