plot.spEMN01 {mixtools}	R Documentation

Plot mixture pdf for the semiparametric mixture model output by spEMsymlocN01

Description

Plot mixture density for the semiparametric mixture model output by spEMsymlocN01, with one component known and set to normal(0,1), and a symmetric nonparametric density with location parameter.

Usage

## S3 method for class 'spEMN01'
plot(x, bw = x$bandwidth, knownpdf = dnorm, add.plot = FALSE, ...)

Arguments

x
An object of class "spEMN01" as returned by spEMsymlocN01

bw
Bandwidth for weighted kernel density estimation.

knownpdf
The known density of component 1, default to dnorm.

add.plot
Set to TRUE to add to an existing plot.

...
further arguments passed to plot if add.plot = FALSE, and to lines if add.plot = TRUE.

Value

A plot of the density of the mixture

Author(s)

Didier Chauveau

References

Chauveau, D., Saby, N., Orton, T. G., Lemercier B., Walter, C. and Arrouys, D. Large-scale simultaneous hypothesis testing in soil monitoring: A semi-parametric mixture approach, preprint (2013).

See Also

spEMsymlocN01

[Package mixtools version 2.0.0.1 Index]