plotly_mixMCMC {mixtools}	R Documentation

Various Plots Pertaining to Mixture Model Output Using MCMC Methods using plotly

Description

This is an updated version of plot.mixMCMC. For technical details, please refer to plot.mixMCMC.

Usage

 
plotly_mixMCMC(x, trace.plot = TRUE, summary.plot = FALSE, burnin = 2000, 
               credit.region = 0.95, col.cr = NULL,
               cex.trace = 3, width.trace = 3, 
               cex.summary = 3, width.summary = 1,
               title.trace = "", title.trace.x = 0.5, 
               title.trace.y = 0.95, title.trace.size = 15,
               xlab.trace = "Index", xlab.trace.size = 15, xtick.trace.size = 15,
               ylab.trace = NULL, ylab.trace.size = 15, ytick.trace.size = 15,
               title.summary = "Credible Regions", title.summary.x = 0.5, 
               title.summary.y = 0.95, title.summary.size = 15,
               xlab.summary = "Predictor", xlab.summary.size = 15, 
               xtick.summary.size = 15,
               ylab.summary = "Response", ylab.summary.size = 15, 
               ytick.summary.size = 15
) 

Arguments

x
An object of class mixMCMC.

trace.plot
If TRUE, trace plots of the various parameters estimated by the MCMC methods is given.

summary.plot
Graphics pertaining to certain mixture models. The details are given below.

burnin
The values 1 to burnin are dropped when producing the plots in summary.plots.

credit.region
Confidence level of credit region.

col.cr
Color of credit region. Number of color specified needs to be consistent with number of components.

cex.trace
Dot size of trace plots.

width.trace
Line width of trace plots.

cex.summary
Dot size of summary plots.

width.summary
Line width of summary plots.

title.trace
Text of the main title of trace plots.

title.trace.x
Horizontal position of main title of trace plots.

title.trace.y
Vertical position of main title of trace plots.

title.trace.size
Text sise of main title of trace plots.

xlab.trace
Label of X-axis of trace plots.

xlab.trace.size
Size of the lable of X-axis of trace plots.

xtick.trace.size
Size of tick lables of X-axis of trace plots.

ylab.trace
Label of Y-axis of trace plots.

ylab.trace.size
Size of the lable of Y-axis of trace plots.

ytick.trace.size
Size of tick lables of Y-axis of trace plots.

title.summary
Text of the main title of summar plot.

title.summary.x
Horizontal position of main title of summary plot.

title.summary.y
Vertical position of main title of summary plot.

title.summary.size
Text sise of main title of summary plot.

xlab.summary
Label of X-axis of summary plot.

xlab.summary.size
Size of the lable of X-axis of summary plot.

xtick.summary.size
Size of tick lables of X-axis of summary plot.

ylab.summary
Label of Y-axis of summary plot.

ylab.summary.size
Size of the lable of Y-axis of summary plot.

ytick.summary.size
Size of tick lables of Y-axis of summary plot.

Value

plotly_mixMCMC returns trace plots of the various parameters estimated by the MCMC methods for all objects of class mixMCMC. In addition, other plots may be produced for the following k-component mixture model functions:

regmixMH
Credible bands for the regression lines in a mixture of linear regressions. See regcr for more details.

See Also

regcr, plot.mixMCMC

Examples

 
## Not run: 
data(NOdata)
attach(NOdata)
set.seed(100)
beta <- matrix(c(1.3, -0.1, 0.6, 0.1), 2, 2)
sigma <- c(.02, .05)
MH.out <- regmixMH(Equivalence, NO, beta = beta, s = sigma,
                   sampsize = 2500, omega = .0013)
plotly_mixMCMC(x = MH.out, summary.plot = TRUE, col.cr = c("red", "green"))

## End(Not run)

[Package mixtools version 2.0.0.1 Index]