plotly_seq.npEM {mixtools}	R Documentation

Plotting sequences of estimates from non- or semiparametric EM-like Algorithm using plotly

Description

This is an updated version of plotseq.npEM. For technical details, please refer to plotseq.npEM.

Usage

  plotly_seq.npEM (x, col = '#1f77b4' , width = 6,
                   xlab = "Iteration" , xlab.size = 15 , xtick.size = 15,
                   ylab.size = 15 , ytick.size = 15,
                   title.size = 15 , title.x = 0.5 , title.y = 0.95)

Arguments

x
an object of class npEM, as output by npEM or spEMsymloc

col
Line color.

width
Line width.

title
Text of the main title.

title.size
Size of the main title.

title.x
Horsizontal position of the main title.

title.y
Vertical posotion of the main title.

xlab
Label of X-axis.

xlab.size
Size of the lable of X-axis.

xtick.size
Size of tick lables of X-axis.

ylab.size
Size of the lable of Y-axis.

ytick.size
Size of tick lables of Y-axis.

Value

plotly_seq.npEM returns a figure with one plot for each component proportion, and, in the case of spEMsymloc, one plot for each component mean.

Author(s)

Didier Chauveau

References

Benaglia, T., Chauveau, D., and Hunter, D. R. (2009), An EM-like algorithm for semi- and non-parametric estimation in multivariate mixtures, Journal of Computational and Graphical Statistics (to appear).

Bordes, L., Chauveau, D., and Vandekerkhove, P. (2007), An EM algorithm for a semiparametric mixture model, Computational Statistics and Data Analysis, 51: 5429-5443.

See Also

plot.npEM, rnormmix, npEM, spEMsymloc, plotly_seq.npEM

Examples

## Not run: 
## Examine and plot water-level task data set.
## First, try a 3-component solution where no two coordinates are
## assumed i.d.
data(Waterdata)
set.seed(100)
## Not run:
a <- npEM(Waterdata[,3:10], mu0=3, bw=4) # Assume indep but not iid
plotly_seq.npEM(a)

## End(Not run)

[Package mixtools version 2.0.0.1 Index]