plotseq.npEM {mixtools}	R Documentation

Plotting sequences of estimates from non- or semiparametric EM-like Algorithm

Description

Returns plots of the sequences of scalar parameter estimates along iterations from an object of class npEM.

Usage

  ## S3 method for class 'npEM'
plotseq(x, ...)

Arguments

x
an object of class npEM, as output by npEM or spEMsymloc

...
further parameters that are passed to plot

Details

plotseq.npEM returns a figure with one plot for each component proportion, and, in the case of spEMsymloc, one plot for each component mean.

References

Benaglia, T., Chauveau, D., and Hunter, D. R. (2009), An EM-like algorithm for semi- and non-parametric estimation in multivariate mixtures, Journal of Computational and Graphical Statistics (to appear).

Bordes, L., Chauveau, D., and Vandekerkhove, P. (2007), An EM algorithm for a semiparametric mixture model, Computational Statistics and Data Analysis, 51: 5429-5443.

See Also

plot.npEM, rnormmix, npEM, spEMsymloc

Examples

## Example from a normal location mixture
n <- 200
set.seed(100)
lambda <- c(1/3,2/3)
mu <- c(0, 4); sigma<-rep(1, 2)
x <- rnormmix(n, lambda, mu, sigma)
b <- spEMsymloc(x, mu0=c(-1, 2), stochastic=FALSE)
plotseq(b)
bst <- spEMsymloc(x, mu0=c(-1, 2), stochastic=TRUE)
plotseq(bst)

[Package mixtools version 2.0.0.1 Index]