plotspRMM {mixtools}	R Documentation

Plot output from Stochastic EM algorithm for semiparametric scaled mixture of censored data

Description

Function for plotting various results from an object returned by spRMM_SEM, a Stochastic EM algorithm for semiparametric scaled mixture of randomly right censored lifetime data. Four plots of sequences of estimates along iterations, survival and density estimates (see reference below).

Usage

  plotspRMM(sem, tmax = NULL)

Arguments

sem
An object returned by spRMM_SEM.

tmax
The max time for x axis, set to some default value if NULL.

Value

The four plots returned

Author(s)

Didier Chauveau

References

Bordes, L., and Chauveau, D. (2016), Stochastic EM algorithms for parametric and semiparametric mixture models for right-censored lifetime data, Computational Statistics, Volume 31, Issue 4, pages 1513-1538. https://link.springer.com/article/10.1007/s00180-016-0661-7

See Also

Related functions: spRMM_SEM.

Other models and algorithms for censored lifetime data (name convention is model_algorithm): expRMM_EM, weibullRMM_SEM.

Examples

# See example(spRMM_SEM)

[Package mixtools version 2.0.0.1 Index]