print.mvnpEM {mixtools}	R Documentation

Printing of Results from the mvnpEM Algorithm Output

Description

print method for class mvnpEM.

Usage

## S3 method for class 'mvnpEM'
print(x, ...)

Arguments

x
an object of class mvnpEM such as a result of a call to mvnpEM

...
Additional arguments to print

Details

print.mvnpEM prints the elements of an mvnpEM object without printing the data or the posterior probabilities. (These may still be accessed as x$data and x$posteriors.)

Value

print.mvnpEM returns (invisibly) the full value of x itself, including the data and posteriors elements.

See Also

mvnpEM, plot.mvnpEM summary.mvnpEM

Examples

# Example as in Chauveau and Hoang (2015) with 6 coordinates
## Not run: 
m=2; r=6; blockid <-c(1,1,2,2,3,3) # 3 bivariate blocks 
# generate some data x ...
a <- mvnpEM(x, mu0=2, blockid, samebw=F) # adaptive bandwidth
print(a)
## End(Not run)

[Package mixtools version 2.0.0.1 Index]