print.npEM {mixtools}	R Documentation

Printing non- and semi-parametric multivariate mixture model fits

Description

print method for class npEM.

Usage

## S3 method for class 'npEM'
print(x, ...)

Arguments

x
an object of class npEM such as a result of a call to npEM

...
Additional arguments to print

Details

print.npEM prints the elements of an npEM object without printing the data or the posterior probabilities. (These may still be accessed as x$data and x$posteriors.)

Value

print.npEM returns (invisibly) the full value of x itself, including the data and posteriors elements.

See Also

npEM, plot.npEM summary.npEM

Examples

data(Waterdata)
set.seed(100)
## Not run: npEM(Waterdata[,3:10], 3, bw=4, verb=FALSE) # Assume indep but not iid

[Package mixtools version 2.0.0.1 Index]