rexpmix {mixtools}	R Documentation

Simulate from Mixtures of Exponentials

Description

Simulate from a mixture of univariate exponential distributions.

Usage

rexpmix(n, lambda = 1, rate = 1)

Arguments

n
Number of cases to simulate.

lambda
Vector of mixture probabilities, with length equal to m, the desired number of components (subpopulations). This is assumed to sum to 1.

rate
Vector of component rates.

Value

rexpmix returns an n-vector sampled from an m-component mixture of univariate exponential distributions.

See Also

rnormmix, rmvnormmix for Gaussian mixtures, rweibullmix for mixture of Weibull distributions.

Examples

## Generate data from a 2-component mixture of exponentials.
n=300 # sample size
m=2 # nb components
lambda=c(1/3, 2/3); rate = c(1,1/10) # parameters
set.seed(1234)
x <- rexpmix(n, lambda, rate) # iid ~ exp mixture

## histogram of the simulated data.
hist(x, col=8)

[Package mixtools version 2.0.0.1 Index]