rmvnorm {mixtools}	R Documentation

Simulate from a Multivariate Normal Distribution

Description

Simulate from a multiviate normal distribution

Usage

rmvnorm(n, mu=NULL, sigma=NULL)

Arguments

n
Number of vectors to simulate

mu
mean vector

sigma
covariance matrix, assumed symmetric and nonnegative definite

Details

This function uses an eigen decomposition assuming sigma is symmetric. In particular, the upper triangle of sigma is ignored.

Value

An n x d matrix in which each row is an independently generated realization from the desired multivariate normal distribution

See Also

eigen, dnorm, dmvnorm

[Package mixtools version 2.0.0.1 Index]