rweibullmix {mixtools}	R Documentation

Simulate from Mixtures of Weibull distributions

Description

Simulate from a mixture of univariate Weibull distributions.

Usage

rweibullmix(n, lambda = 1, shape = 1, scale = 1)

Arguments

n
Number of cases to simulate.

lambda
Vector of mixture probabilities, with length equal to m, the desired number of components (subpopulations). This is assumed to sum to 1.

shape
Vector of component shapes.

scale
Vector of component scales.

Value

rexpmix returns an n-vector sampled from an m-component mixture of univariate Weibull distributions.

See Also

rnormmix and rmvnormmix for Gaussian mixtures, rexpmix for mixture of exponentials.

Examples

n = 500 # sample size
m = 2 # nb components
lambda=c(0.4, 0.6)
shape <- c(0.5,5); scale <- c(1,20) # model parameters
set.seed(321)
x <- rweibullmix(n, lambda, shape, scale) # iid ~ weibull mixture

## histogram of the simulated data.
hist(x, col=8)

[Package mixtools version 2.0.0.1 Index]