summary.mixEM {mixtools}	R Documentation

Summarizing EM mixture model fits

Description

summary method for class mixEM.

Usage

## S3 method for class 'mixEM'
summary(object, digits=6, ...)

Arguments

object
an object of class mixEM such as a result of a call to normalmixEM

digits
Significant digits for printing values

...
further arguments passed to print method.

Details

summary.mixEM prints parameter estimates for each component of a fitted mixture model. The estimates printed vary with the type of model.

Value

The function summary.mixEM prints the final loglikelihood value at the solution as well as a matrix of values for each component that could include:

lambda
The estimated mixing weights

mu
The estimated mean parameters

sigma
The estimated standard deviations

theta
The estimated multinomial parameters

beta
The estimated regression parameters

See Also

normalmixEM, logisregmixEM, multmixEM, mvnormalmixEM, poisregmixEM, regmixEM, regmixEM.lambda, regmixEM.loc, regmixEM.mixed, regmixEM.chgpt, repnormmixEM, expRMM_EM, weibullRMM_SEM

Examples

data(faithful)
attach(faithful)
set.seed(100)
out <- normalmixEM(waiting, mu=c(50,80), sigma=c(5,5), lambda=c(.5,.5))
summary(out)

[Package mixtools version 2.0.0.1 Index]