summary.mvnpEM {mixtools}	R Documentation

Summarizing Fits for Nonparametric Mixture Models with Conditionally Independent Multivariate Component Densities

Description

summary method for class mvnpEM.

Usage

## S3 method for class 'mvnpEM'
summary(object, ...)
## S3 method for class 'summary.mvnpEM'
print(x, digits=3, ...)

Arguments

object,x
an object of class mvnpEM such as a result of a call to mvnpEM

digits
Significant digits for printing values

...
further arguments passed to or from other methods.

Details

summary.mvnpEM prints means and variances of each block for each component. These quantities might not be part of the model, but they are estimated nonparametrically based on the posterior probabilities and the data.

Value

The function summary.mvnpEM returns a list of type summary.mvnpEM with the following components:

n
The number of observations

m
The number of mixture components

B
The number of blocks

blockid
The block ID (from 1 through B) for each of the coordinates of the multivariate observations. The blockid component is of length r, the dimension of each observation.

means
A B x m matrix giving the estimated mean of each block in each component.

variances
Same as means but giving the estimated variances instead.

References

Benaglia, T., Chauveau, D., and Hunter, D. R. (2009), An EM-like algorithm for semi- and non-parametric estimation in multivariate mixtures, Journal of Computational and Graphical Statistics, 18(2), 505–526.

Chauveau, D., and Hoang, V. T. L. (2015), Nonparametric mixture models with conditionally independent multivariate component densities, Preprint under revision. https://hal.science/hal-01094837

See Also

mvnpEM, plot.mvnpEM

Examples

# Example as in Chauveau and Hoang (2015) with 6 coordinates
## Not run: 
m=2; r=6; blockid <-c(1,1,2,2,3,3) # 3 bivariate blocks 
# generate some data x ...
a <- mvnpEM(x, mu0=2, blockid, samebw=F) # adaptive bandwidth
plot(a) # this S3 method produces 6 plots of univariate marginals
summary(a)
## End(Not run)

[Package mixtools version 2.0.0.1 Index]