summary.npEM {mixtools}	R Documentation

Summarizing non- and semi-parametric multivariate mixture model fits

Description

summary method for class npEM.

Usage

## S3 method for class 'npEM'
summary(object, ...)
## S3 method for class 'summary.npEM'
print(x, digits=3, ...)

Arguments

object,x
an object of class npEM such as a result of a call to npEM

digits
Significant digits for printing values

...
further arguments passed to or from other methods.

Details

summary.npEM prints means and variances of each block for each component. These quantities might not be part of the model, but they are estimated nonparametrically based on the posterior probabilities and the data.

Value

The function summary.npEM returns a list of type summary.npEM with the following components:

n
The number of observations

m
The number of mixture components

B
The number of blocks

blockid
The block ID (from 1 through B) for each of the coordinates of the multivariate observations. The blockid component is of length r, the dimension of each observation.

means
A B x m matrix giving the estimated mean of each block in each component.

variances
Same as means but giving the estimated variances instead.

References

Benaglia, T., Chauveau, D., and Hunter, D. R. (2009), An EM-like algorithm for semi- and non-parametric estimation in multivariate mixtures, Journal of Computational and Graphical Statistics, 18(2), 505–526.

See Also

npEM, plot.npEM

Examples

data(Waterdata)
set.seed(100)
## Not run: 
a <- npEM(Waterdata[,3:10], 3, bw=4) # Assume indep but not iid
summary(a) 

b <- npEM(Waterdata[,3:10], 3, bw=4, blockid=rep(1,8)) # Now assume iid
summary(b)

## End(Not run)

[Package mixtools version 2.0.0.1 Index]