summary.spRMM {mixtools}	R Documentation

Summarizing fits from Stochastic EM algorithm for semiparametric scaled mixture of censored data

Description

summary method for class spRMM.

Usage

## S3 method for class 'spRMM'
summary(object, digits = 6, ...)

Arguments

object
an object of class spRMM such as a result of a call to spRMM_SEM

digits
Significant digits for printing values

...
Additional parameters passed to print.

Details

summary.spRMM prints scalar parameter estimates for a fitted mixture model: each component weight and the scaling factor, see reference below. The functional (nonparametric) estimates of survival and hazard rate funcions can be obtained using plotspRMM.

Value

The function summary.spRMM prints the final loglikelihood value at the solution as well as The estimated mixing weights and the scaling parameter.

Author(s)

Didier Chauveau

References

Bordes, L., and Chauveau, D. (2016), Stochastic EM algorithms for parametric and semiparametric mixture models for right-censored lifetime data, Computational Statistics, Volume 31, Issue 4, pages 1513-1538. https://link.springer.com/article/10.1007/s00180-016-0661-7

See Also

Function for plotting functional (nonparametric) estimates: plotspRMM.

Other models and algorithms for censored lifetime data (name convention is model_algorithm): expRMM_EM, weibullRMM_SEM.

Examples

# See example(spRMM_SEM)

[Package mixtools version 2.0.0.1 Index]